requirements.txt
requests
python-dotenv
pandas
psycopg2-binary
sqlalchemy
apache-airflow